+++ pkg/framework/objects/scim_group_partial_permissions/model.go
@@ -57,7 +57,7 @@ func CompareScimGroupPartialPermissions(
 func ConvertScimGroupPartialPermissionModelToData(
 	requiredAllPermissions []ScimGroupPartialPermissionModel,
 	groupID int,
-	accountID int,
+	accountID int64,
 ) []dbt_cloud.GroupPermission {
 	allPermissions := []dbt_cloud.GroupPermission{}
